Split playback span helpers
